pushing the functions form cliCSD to module
