+++ cliCSD.py
@@ -6,7 +6,7 @@ The idea is to be able to use the saved geometry file
 and deliver the required input as a command line
 parameters.
 
-As an output the myLoged info of power losses
+As an output the csdf.myLoged info of power losses
 is generated on the standard output.
 """
 
@@ -16,91 +16,26 @@ is generated on the standard output.
 # 3. Setup the solver - done
 # 4. Solve - done
 # 5. Prepare results - done
-# 6. myLog results - done
+# 6. csdf.myLog results - done
 # 7. adding inner code working - done
+# 8. cleanu and make use of modules - done
 
 # General imports
 import numpy as np
 import os.path
 import sys
-
 import pickle
-
-# 1.
 import argparse
 
-
-# global constant to speed up loops
-mi0 = 4 * np.pi * 1e-7
-C1 = -np.log(2) / 3 + 13 / 12 - np.pi / 2
-C2 = mi0 / (2 * np.pi)
-
-# # Importing local library
-# from csdlib import csdlib as csd
-from csdlib import innercode as ic
+# Importing local library
 from csdlib import csdfunctions as csdf
-
-
-# making the NUMBA decorators optional
-def conditional_decorator(dec, condition):
-    def decorator(func):
-        if not condition:
-            # Return the function unchanged, not decorated.
-            return func
-        return dec(func)
-
-    return decorator
-
-
-try:
-    from numba import njit
-
-    use_njit = True
-except ImportError:
-    use_njit = False
-    njit = None
-
-# use_njit = not True
-
-
-# 2
-def loadTheData(filename):
-    """
-    This is sub function to load data
-    """
-
-    if os.path.isfile(filename):
-        _, extension = os.path.splitext(filename)
-        myLog("File type: " + extension)
-
-        if extension.lower() in [".txt", ".inc", ".ic"]:
-            myLog("reading the inner-code geometry file: " + filename)
-            try:
-                with open(filename, "r") as f:
-                    file_content = f.read()
-
-                XSecArray, dXmm, dYmm = getCanvas(file_content)
-            except IOError:
-                print("Error reading the file " + filename)
-                sys.exit(1)
-
-        else:
-            myLog("reading from file :" + filename)
-            XSecArray, dXmm, dYmm = loadObj(filename).restore()
-
-        return XSecArray, dXmm, dYmm
-    else:
-        myLog(f"The file {filename} can't be opened!")
-        sys.exit(1)
-
-
-def myLog(s: str = "", *args, **kwargs):
-    if verbose:
-        print(s, *args, *kwargs)
+from csdlib import csdmath as csdm
 
 
 def getArgs():
-    # 1 handling the in line parameters
+    """
+    Handling the cli line parameters. 
+    """
     parser = argparse.ArgumentParser(
         description="CSD cli executor",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
@@ -152,407 +87,22 @@ def getArgs():
     return vars(args)
 
 
-def loadObj(filename):
-    """load object data from file that was saved by saveObj function.
-    Inputs:
-    filename: file to save the data to with properly delivered path.
-
-    Returns:
-    Recreated object
-
-    Example use:``
-
-    P = loadObj('project.save')
-    recreate P as myProject class object.
+def main():
     """
-    with open(filename, "rb") as myInput:
-        return pickle.load(myInput)
-
-
-def getCanvas(textInput):
-    """This function is to determine the best parameters for the canvas
-    based on the given geometry steps defined by the inner code."""
-
-    global XSecArray, dXmm, dYmm
-
-    codeLines = textInput.splitlines()
-    codeSteps = ic.textToCode(codeLines)
-
-    X = []
-    Y = []
-
-    circles = False
-    if codeSteps:
-        for step in codeSteps:
-            tmp = step[0](*step[1], draw=False)
-            if step[0] is ic.addCircle:
-                circles = True
-
-            X.append(tmp[0])
-            X.append(tmp[2])
-            Y.append(tmp[1])
-            Y.append(tmp[3])
-
-        myLog(X)
-        myLog(Y)
-        myLog(f"Dimention range: {min(X)}:{max(X)}; {min(Y)}:{max(Y)}")
-        size = (max(X) - min(X), max(Y) - min(Y))
-        myLog(size)
-
-        # I have no good idea how to figure out the best cell size
-        # so for now it's just some stuff..
-        if circles:
-            sizes = [4, 2.5, 2, 1]
-        else:
-            sizes = [15, 10, 5, 4, 2.5, 2, 1]
-        for xd in sizes:
-            if (size[0] % xd == 0) and (size[1] % xd == 0):
-                break
-        myLog(f"The dx: {xd}mm")
-
-        elements_x = int(size[1] / xd)
-        elements_y = int(size[0] / xd)
-
-        myLog(f"Canvas elements need: {elements_x, elements_y}")
-
-        dXmm = dYmm = xd
-        XSecArray = np.zeros([elements_x, elements_y])
-
-        for step in codeSteps:
-            step[0](*step[1], shift=(min(X), min(Y)), XSecArray=XSecArray, dXmm=dXmm)
-        return XSecArray, dXmm, dYmm
-
-    return False
-
-
-def solveTheEquation(admitanceMatrix, voltageVector):
-    return np.matmul(admitanceMatrix, voltageVector)
-
-
-def getGmatrix(input):
-    return np.linalg.inv(input)
-
-
-# @njit
-@conditional_decorator(njit, use_njit)
-def N_getImpedanceArray(
-    distanceArray,
-    freq,
-    dXmm,
-    dYmm,
-    lenght=1000,
-    temperature=20,
-    sigma20C=58e6,
-    temCoRe=3.9e-3,
-):
-    """
-    Calculate the array of impedance as complex values for each element
-    Input:
-    distanceArray -  array of distances beetween the elements in [mm]
-    freq = frequency in Hz
-    dXmm - size of element in x [mm]
-    dYmm - size of element in Y [mm]
-    lenght - analyzed lenght in [mm] /default= 1000mm
-    temperature - temperature of the conductors in deg C / default = 20degC
-    sigma20C - conductivity of conductor material in 20degC in [S] / default = 58MS (copper)
-    temCoRe - temperature resistance coefficient / default is copper
-    """
-    omega = 2 * np.pi * freq
-
-    impedanceArray = np.zeros((distanceArray.shape), dtype=np.complex128)
-
-    for X in range(distanceArray.shape[0]):
-        for Y in range(distanceArray.shape[0]):
-            if X == Y:
-                impedanceArray[Y, X] = N_getResistance(
-                    sizeX=dXmm,
-                    sizeY=dYmm,
-                    lenght=lenght,
-                    temp=temperature,
-                    sigma20C=sigma20C,
-                    temCoRe=temCoRe,
-                ) + 1j * omega * N_getSelfInductance(
-                    sizeX=dXmm, sizeY=dYmm, lenght=lenght
-                )
-            else:
-                impedanceArray[Y, X] = (
-                    1j
-                    * omega
-                    * N_getMutualInductance(
-                        sizeX=dXmm,
-                        sizeY=dYmm,
-                        lenght=lenght,
-                        distance=distanceArray[Y, X],
-                    )
-                )
-    return impedanceArray
-
-
-# Calculation of self inductance value function
-# @njit
-@conditional_decorator(njit, use_njit)
-def N_getSelfInductance(sizeX, sizeY, lenght):
-    """
-    Calculate the self inductance for the subconductor
-    It assumes rectangular shape. If you want put for circular shape just
-    make sizeX = sizeY = 2r
-
-    Inputs:
-    sizeX - width in [mm]
-    sizeY - height in [mm]
-    lenght - cinductor lenght in [mm]
-
-    output
-    L in [H]
-    """
-    srednica = (sizeX + sizeY) / 2
-    a = srednica * 1e-3
-    l = lenght * 1e-3
-
-    # This calculation is based on the:
-    # https://pdfs.semanticscholar.org/b0f4/eff92e31d4c5ff42af4a873ebdd826e610f5.pdf
-    # L = (mi0 * l / (2 * np.pi)) * (
-    #     np.log(2 * l / a) - np.log(2) / 3 + 13 / 12 - np.pi / 2
-    # )
-
-    # this is the above formula just with pre calculated constant
-    L = (C2 * l) * (np.log(2 * l / a) + C1)
-
-    # this was the previous formula
-    # return 0.000000001*2*100*lenght*1e-3*(np.log(2*lenght*1e-3/(0.5*srednica*1e-3))-(3/4))
-
-    return L
-
-
-# @njit
-@conditional_decorator(njit, use_njit)
-def N_getResistance(sizeX, sizeY, lenght, temp, sigma20C, temCoRe):
-    """
-    Calculate the resistance of the al'a square shape in given temperature
-    All dimensions in mm
-    temperature in deg C
-
-    output:
-    Resistance in Ohm
-    """
-    return (lenght / (sizeX * sizeY * sigma20C)) * 1e3 * (1 + temCoRe * (temp - 20))
-
-
-# Calculations of mutual inductance between conductors
-# @njit
-@conditional_decorator(njit, use_njit)
-def N_getMutualInductance(sizeX, sizeY, lenght, distance):
-    """
-    Calculate the mutual inductance for the subconductor
-    It assumes rectangular shape. If you want put for circular shape just
-    make sizeX = sizeY = 2r
-
-    Inputs:
-    sizeX - width in [mm]
-    sizeY - height in [mm]
-    lenght - conductor lenght in [mm]
-    distance - distance between analyzed conductors in [mm]
-
-    output
-    M in [H]
-    """
-    srednica = (sizeX + sizeY) / 2
-
-    a = srednica * 1e-3
-    l = lenght * 1e-3
-    d = distance * 1e-3
-    # mi0 = 4 * np.pi * 1e-7
-
-    # formula by:
-    # https://pdfs.semanticscholar.org/b0f4/eff92e31d4c5ff42af4a873ebdd826e610f5.pdf
-    # M = (mi0 * l / (2 * np.pi)) * (
-    #     np.log((l + np.sqrt(l**2 + d**2)) / d) - np.sqrt(1 + (d / l) ** 2) + d / l
-    # )
-    # the same as above with the pre-calculated C2
-    M = (
-        C2
-        * l
-        * (np.log((l + np.sqrt(l**2 + d**2)) / d) - np.sqrt(1 + (d / l) ** 2) + d / l)
-    )
-
-    # previous formula
-    # return 0.000000001*2*lenght*1e-1*(np.log(2*lenght*1e-1/(distance/10))-(3/4))
-    return M
-
-
-# @njit
-@conditional_decorator(njit, use_njit)
-def N_getResistanceArray(
-    elementsVector,
-    dXmm,
-    dYmm,
-    lenght=1000,
-    temperature=20,
-    sigma20C=58e6,
-    temCoRe=3.9e-3,
-):
-    """
-    Calculate the array of resistance values for each element
-    Input:
-    elementsVector - The elements vector as delivered by arrayVectorize
-    dXmm - size of element in x [mm]
-    dYmm - size of element in Y [mm]
-    lenght - analyzed lenght in [mm] /default= 1000mm
-    temperature - temperature of the conductors in deg C / defoult = 20degC
-    sigma20C - conductivity of conductor material in 20degC in [S] / default = 58MS (copper)
-    temCoRe - temperature resistance coeficcient / default is copper
+    This is the place where the main flow of operation is carried.
     """
 
-    resistanceArray = np.zeros(elementsVector.shape[0])
-    for element in range(elementsVector.shape[0]):
-        resistanceArray[element] = N_getResistance(
-            sizeX=dXmm,
-            sizeY=dYmm,
-            lenght=lenght,
-            temp=temperature,
-            sigma20C=sigma20C,
-            temCoRe=temCoRe,
-        )
-
-    return resistanceArray
-
-
-# Calculate distance between elements function
-# @njit
-@conditional_decorator(njit, use_njit)
-def N_getDistancesArray(inputVector):
-    """
-    This function calculate the array of distances between every conductors
-    element
-    Input:
-    the vector of conductor elements as delivered by n_vectorizeTheArray
-    """
-    # lets check for the numbers of elements
-    elements = inputVector.shape[0]
-    # print(elements)
-    # Define the outpur array
-    distanceArray = np.zeros((elements, elements))
-
-    for x in range(elements):
-        for y in range(elements):
-            if x != y:
-                posXa = inputVector[y][2]
-                posYa = inputVector[y][3]
-
-                posXb = inputVector[x][2]
-                posYb = inputVector[x][3]
-
-                distanceArray[y, x] = np.sqrt(
-                    (posXa - posXb) ** 2 + (posYa - posYb) ** 2
-                )
-            else:
-                distanceArray[y, x] = 0
-    return distanceArray
-
-
-# Master Array Vectorization FUNCTION
-# @njit
-# @conditional_decorator(njit, use_njit) # there is an issue with numba here if the output array is empty one. ``
-def N_arrayVectorize(inputArray, phaseNumber, dXmm, dYmm):
-    """
-    Desription:
-    This function returns vector of 4 dimension vectors that deliver
-
-    input:
-    inputArray = 3D array thet describe by 1's position of
-    conductors in cross section
-    dXmm - size of each element in X direction [mm]
-    dYmm - size of each element in Y direction [mm]
-    Output:
-    [0,1,2,3] - 4 elements vector for each element, where:
-
-    0 - Original inputArray geometry origin Row for the set cell
-    1 - Original inputArray geometry origin Col for the set cell
-    2 - X position in mm of the current element
-    3 - Y position in mm of the current element
-
-    Number of such [0,1,2,3] elements is equal to the number of defined
-    conductor cells in geometry
-
-    """
-    # Let's check the size of the array
-    elementsInY = inputArray.shape[0]
-    elementsInX = inputArray.shape[1]
-
-    # lets define the empty vectorArray
-    vectorArray = []
-
-    # lets go for each input array position and check if is set
-    # and if yes then put it into putput vectorArray
-    for Row in range(elementsInY):
-        for Col in range(elementsInX):
-            if inputArray[Row][Col] == phaseNumber:
-                # Let's calculate the X and Y coordinates
-                coordinateY = (0.5 + Row) * dYmm
-                coordinateX = (0.5 + Col) * dXmm
-
-                vectorArray.append([Row, Col, coordinateX, coordinateY])
-
-    return np.array(vectorArray)
-
-
-# @njit
-# @conditional_decorator(njit, use_njit)
-def N_arraySlicer(inputArray, subDivisions=2):
-    """
-    This function increase the resolution of the cross section array
-    inputArray -  oryginal geometry matrix
-    subDivisions -  number of subdivisions / factor of increase of resoluttion / default = 2
-    """
-    return inputArray.repeat(subDivisions, axis=0).repeat(subDivisions, axis=1)
-
-
-def N_getComplexModule(x):
-    """
-    returns the module of complex number
-    input: x - complex number
-    if not a complex number is given as parameter then it return the x diretly
-
-    """
-    if isinstance(x, complex):
-        return np.absolute(x)
-        # return np.sqrt(x.real ** 2 + x.imag ** 2)
-    else:
-        return x
-
-
-# Function that put back together the solution vectr back to represent the cross
-# section shape array
-def N_recreateresultsArray(elementsVector, resultsVector, initialGeometryArray):
-    """
-    Functions returns recreate cross section array with mapperd solution results
-    Inputs:
-    elementsVector - vector of crossection elements as created by the n_arrayVectorize
-    resultsVector - vectr with results values calculated base on the elementsVector
-    initialGeometryArray - the array that contains the cross section geometry model
-    """
-    localResultsArray = np.zeros((initialGeometryArray.shape), dtype=float)
-
-    for vectorIndex, result in enumerate(resultsVector):
-        localResultsArray[
-            int(elementsVector[vectorIndex][0]), int(elementsVector[vectorIndex][1])
-        ] = result
-
-    return localResultsArray
-
-
-
-
-# Doing the main work here.
-if __name__ == "__main__":
     config = getArgs()
     verbose = config["verbose"]
     simple = config["simple"]
     csv = config["csv"]
 
-    myLog()
-    myLog("Starting operations...")
-    myLog()
+    # for simplicity so the log procrdure can see it globally
+    csdf.verbose = verbose
+
+    csdf.myLog()
+    csdf.myLog("Starting operations...")
+    csdf.myLog()
 
     if config["draw"] or config["results"]:
         import matplotlib.pyplot as plt
@@ -562,21 +112,22 @@ if __name__ == "__main__":
     dXmm = dYmm = 1
 
     # 2 loading the geometry data:
-    XSecArray, dXmm, dYmm = loadTheData(config["geometry"])
-    myLog("Initial geometry array parameters:")
-    myLog(f"dX:{dXmm}mm dY:{dYmm}mm")
-    myLog(f"Data table size: {XSecArray.shape}")
+    XSecArray, dXmm, dYmm = csdf.loadTheData(config["geometry"])
+
+    csdf.myLog("Initial geometry array parameters:")
+    csdf.myLog(f"dX:{dXmm}mm dY:{dYmm}mm")
+    csdf.myLog(f"Data table size: {XSecArray.shape}")
 
     while dXmm > config["size"]:
-        myLog("Splitting the geometry cells...", end="")
-        XSecArray = N_arraySlicer(inputArray=XSecArray, subDivisions=2)
+        csdf.myLog("Splitting the geometry cells...", end="")
+        XSecArray = csdm.arraySlicer(inputArray=XSecArray, subDivisions=2)
         dXmm = dXmm / 2
         dYmm = dYmm / 2
 
-    myLog()
-    myLog("Adjusted geometry array parameters:")
-    myLog(f"dX:{dXmm}mm dY:{dYmm}mm")
-    myLog(f"Data table size: {XSecArray.shape}")
+    csdf.myLog()
+    csdf.myLog("Adjusted geometry array parameters:")
+    csdf.myLog(f"dX:{dXmm}mm dY:{dYmm}mm")
+    csdf.myLog(f"Data table size: {XSecArray.shape}")
 
     if config["draw"]:
         # making the draw of the geometry in initial state.
@@ -614,53 +165,54 @@ if __name__ == "__main__":
     # 3 preparing the solution
     Irms = config["current"]
     # Current vector
-    I = [Irms, 0, Irms, 120, Irms, 240]
+    I = [Irms, 120, Irms, 0, Irms, 240]
     f = config["frequency"]
     length = config["length"]
     t = config["Temperature"]
 
-    myLog()
-    myLog("Starting solver for")
-    for k, n in zip([0, 2, 4], ["a", "b", "c"]):
-        myLog(f"I{n} = {I[k]}[A] \t {I[k+1]}[deg] \t {f}[Hz]")
+    csdf.myLog()
+    csdf.myLog("Starting solver for")
 
-    myLog()
-    myLog("Complex form:")
+    for k, n in zip([0, 2, 4], ["a", "b", "c"]):
+        csdf.myLog(f"I{n} = {I[k]}[A] \t {I[k+1]}[deg] \t {f}[Hz]")
+
+    csdf.myLog()
+    csdf.myLog("Complex form:")
 
     # lets workout the  current in phases as is defined
     in_Ia = I[0] * np.cos(I[1] * np.pi / 180) + I[0] * np.sin(I[1] * np.pi / 180) * 1j
-    myLog(f"Ia: {in_Ia}")
+    csdf.myLog(f"Ia: {in_Ia}")
 
     in_Ib = I[2] * np.cos(I[3] * np.pi / 180) + I[2] * np.sin(I[3] * np.pi / 180) * 1j
-    myLog(f"Ib: {in_Ib}")
+    csdf.myLog(f"Ib: {in_Ib}")
 
     in_Ic = I[4] * np.cos(I[5] * np.pi / 180) + I[4] * np.sin(I[5] * np.pi / 180) * 1j
-    myLog(f"Ic: {in_Ic}")
+    csdf.myLog(f"Ic: {in_Ic}")
 
-    vPhA = N_arrayVectorize(inputArray=XSecArray, phaseNumber=1, dXmm=dXmm, dYmm=dYmm)
-    vPhB = N_arrayVectorize(inputArray=XSecArray, phaseNumber=2, dXmm=dXmm, dYmm=dYmm)
-    vPhC = N_arrayVectorize(inputArray=XSecArray, phaseNumber=3, dXmm=dXmm, dYmm=dYmm)
+    vPhA = csdm.arrayVectorize(inputArray=XSecArray, phaseNumber=1, dXmm=dXmm, dYmm=dYmm)
+    vPhB = csdm.arrayVectorize(inputArray=XSecArray, phaseNumber=2, dXmm=dXmm, dYmm=dYmm)
+    vPhC = csdm.arrayVectorize(inputArray=XSecArray, phaseNumber=3, dXmm=dXmm, dYmm=dYmm)
 
     elementsVector, elementsPhaseA, elementsPhaseB, elementsPhaseC = csdf.combineVectors(
         vPhA, vPhB, vPhC
     )
 
     if len(elementsVector) > 1200:
-        myLog()
-        myLog(
+        csdf.myLog()
+        csdf.myLog(
             "!!! Size of the elements vector may lead to very long calculation. Be aware!"
         )
-        myLog("You can break the process by CTRL+C")
-        myLog("You may conceder reduce the split steps.")
-        myLog("Optimal element size is around 1.5x1.5mm")
-        myLog()
+        csdf.myLog("You can break the process by CTRL+C")
+        csdf.myLog("You may conceder reduce the split steps.")
+        csdf.myLog("Optimal element size is around 1.5x1.5mm")
+        csdf.myLog()
 
     if len(elementsVector) > 10000:
-        myLog("Extreme size of elements vector - long calculations immanent!")
+        csdf.myLog("Extreme size of elements vector - long calculations immanent!")
 
-    admitanceMatrix = getGmatrix(
-        N_getImpedanceArray(
-            N_getDistancesArray(elementsVector),
+    admitanceMatrix = csdm.getGmatrix(
+        csdm.getImpedanceArray(
+            csdm.getDistancesArray(elementsVector),
             freq=f,
             dXmm=dXmm,
             dYmm=dYmm,
@@ -683,7 +235,7 @@ if __name__ == "__main__":
 
     # Initial solve
     # Main equation solve
-    currentVector = solveTheEquation(admitanceMatrix, voltageVector)
+    currentVector = csdm.solveTheEquation(admitanceMatrix, voltageVector)
 
     # And now we need to get solution for each phase to normalize it
     currentPhA = currentVector[0:elementsPhaseA]
@@ -705,11 +257,11 @@ if __name__ == "__main__":
     if Ic:
         Uc = Uc * (in_Ic / Ic)
 
-    myLog()
-    myLog("Calculated require Source Voltages")
-    myLog(Ua)
-    myLog(Ub)
-    myLog(Uc)
+    csdf.myLog()
+    csdf.myLog("Calculated require Source Voltages")
+    csdf.myLog(Ua)
+    csdf.myLog(Ub)
+    csdf.myLog(Uc)
 
     # Setting up the voltage vector for final solve
     vA = np.ones(elementsPhaseA) * Ua
@@ -720,7 +272,7 @@ if __name__ == "__main__":
 
     # Final solve
     # Main equation solve
-    currentVector = solveTheEquation(admitanceMatrix, voltageVector)
+    currentVector = csdm.solveTheEquation(admitanceMatrix, voltageVector)
 
     # And now we need to get solution for each phase to normalize it
     currentPhA = currentVector[0:elementsPhaseA]
@@ -733,14 +285,14 @@ if __name__ == "__main__":
     Ic = np.sum(currentPhC)
     # end of second solve!
 
-    myLog()
-    myLog("Solution check...")
-    myLog("Raw Current results:")
-    myLog(f"Ia: {Ia}")
-    myLog(f"Ib: {Ib}")
-    myLog(f"Ic: {Ic}")
-    myLog()
-    myLog(f"Sum: {Ia+Ib+Ic}")
+    csdf.myLog()
+    csdf.myLog("Solution check...")
+    csdf.myLog("Raw Current results:")
+    csdf.myLog(f"Ia: {Ia}")
+    csdf.myLog(f"Ib: {Ib}")
+    csdf.myLog(f"Ic: {Ic}")
+    csdf.myLog()
+    csdf.myLog(f"Sum: {Ia+Ib+Ic}")
 
     # Now we normalize up to the expecter I - just a polish
     # as we are almost there with the previous second solve for new VOLTAGES
@@ -759,22 +311,22 @@ if __name__ == "__main__":
     Ib = np.sum(currentPhB)
     Ic = np.sum(currentPhC)
 
-    myLog("Fix Current results:")
-    myLog(f"Ia: {Ia}")
-    myLog(f"Ib: {Ib}")
-    myLog(f"Ic: {Ic}")
-    myLog()
-    myLog(f"Sum: {Ia+Ib+Ic}")
+    csdf.myLog("Fix Current results:")
+    csdf.myLog(f"Ia: {Ia}")
+    csdf.myLog(f"Ib: {Ib}")
+    csdf.myLog(f"Ic: {Ic}")
+    csdf.myLog()
+    csdf.myLog(f"Sum: {Ia+Ib+Ic}")
 
     # Data postprocessing
-    getMod = np.vectorize(N_getComplexModule)
+    getMod = np.vectorize(csdm.getComplexModule)
 
     resultsCurrentVector = np.concatenate((currentPhA, currentPhB, currentPhC), axis=0)
     # for debug
-    # myLog(resultsCurrentVector)
+    # csdf.myLog(resultsCurrentVector)
     #
     resultsCurrentVector = getMod(resultsCurrentVector)
-    resistanceVector = N_getResistanceArray(
+    resistanceVector = csdm.getResistanceArray(
         elementsVector, dXmm=dXmm, dYmm=dYmm, temperature=t, lenght=length
     )
 
@@ -809,7 +361,7 @@ if __name__ == "__main__":
     if config["results"]:
         # getting the current density
         resultsCurrentVector *= 1 / (dXmm * dYmm)
-        currentsDraw = N_recreateresultsArray(
+        currentsDraw = csdm.recreateresultsArray(
             elementsVector, resultsCurrentVector, XSecArray
         )
         minCurrent = resultsCurrentVector.min()
@@ -857,3 +409,7 @@ Current Density distribution [A/mm2]",
         )
 
         plt.show()
+
+# Doing the main work here.
+if __name__ == "__main__":
+    main()
